+++ sistema_petshop/src/main/java/com/petshop/services/EspecieService.java
@@ -1,43 +1,51 @@
 package com.petshop.services;
 
-import java.util.List;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
 import com.petshop.model.Especie;
 import com.petshop.repository.EspecieRepository;
 
 import jakarta.persistence.EntityNotFoundException;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.Optional;
 
 @Service
 public class EspecieService {
 
-    @Autowired
-    private EspecieRepository especieRepository;
+    private final EspecieRepository especieRepository;
+
+    public EspecieService(EspecieRepository especieRepository) {
+        this.especieRepository = especieRepository;
+    }
 
     public List<Especie> buscarTodasAsEspecies() {
         return especieRepository.findAll();
     }
 
-    public void salvarEspecie(Especie especie) {
-        especieRepository.save(especie);
+    public Optional<Especie> buscarPorId(Integer id) {
+        return especieRepository.findById(id);
     }
 
-    public Especie buscarPorId(Integer id) {
-        return especieRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Vendedor não encontrado com ID: " + id));
+    public Especie buscarPorIdOuFalhar(Integer id) {
+        return buscarPorId(id)
+                .orElseThrow(() -> new EntityNotFoundException("Espécie não encontrada com ID: " + id));
+    }
+
+    public Especie salvarEspecie(Especie especie) {
+        if (especie.getId() != null) {
+            Especie existente = buscarPorIdOuFalhar(especie.getId());
+            existente.setNome(especie.getNome());
+            return especieRepository.save(existente);
+        } else {
+            return especieRepository.save(especie);
+        }
     }
 
     public void excluirEspeciePorId(Integer id) {
+        if (!especieRepository.existsById(id)) {
+            throw new EntityNotFoundException("Espécie não encontrada com ID: " + id);
+        }
         especieRepository.deleteById(id);
     }
-
-    public Especie atualizarEspecie(Especie especie) {
-        if (especie.getId() != null) {
-            return especieRepository.save(especie);
-        }
-        return null;
-    }
-
 }
